requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
google-generativeai==0.4.1
streamlit==1.32.0
requests==2.31.0
streamlit-folium
folium
polyline
plotly
pandas
